+++ src/core/session.ts
@@ -1,0 +1,137 @@
+export class TimeLogSession implements ITimeLog {
+    active: IActive;
+    beg: TTimestamp;
+    end: TTimestamp;
+    elapsed: number;
+    procedure: IInterval[];
+    memo: string;
+
+    timer: any;
+    currentInterval: IInterval;
+    runningElapsed: number;
+    runId: string;
+
+    status: "running" | "pause" | "stop";
+
+    callbacks = [];
+
+    constructor(active: IActive) {
+        this.active = active;
+        this.procedure = [];
+        this.beg = null;
+        this.end = null;
+        this.elapsed = 0;
+        this.memo = "";
+        this.runningElapsed = 0;
+        this.callbacks = [];
+        this.status = "pause";
+    }
+
+    export(): ITimeLog {
+        return {
+            active: this.active,
+            beg: this.beg,
+            end: this.end,
+            procedure: this.procedure,
+            elapsed: this.elapsed,
+            memo: this.memo,
+        }
+    }
+
+    addCallback(callback: Function) {
+        this.callbacks.push(callback);
+    }
+
+    /**
+     * 供 svelte 组件重载, 以方便触发更新
+     */
+    updateActiveCallback() {
+
+    }
+
+    start() {
+        // console.log("start");
+        let startTime = Date.now();
+        this.beg = startTime;
+        let intervale: IInterval = {
+            beg: startTime,
+            end: -1,
+            elapsed: 0,
+        };
+        this.procedure.push(intervale);
+
+        this.currentInterval = intervale;
+        this.timer = setInterval(() => {
+            let elapsed = Date.now() - this.currentInterval.beg;
+            this.currentInterval.elapsed = elapsed;
+            this.runningElapsed = this.elapsed + elapsed;
+            // console.log(this.runningElapsed);
+            this.callbacks.forEach(callback => {
+                callback(this.runningElapsed);
+            });
+        }, 1000);
+    }
+
+    pause() {
+        if (!this.currentInterval) return;
+        let endTime = Date.now();
+        this.currentInterval.end = endTime;
+        this.currentInterval.elapsed = endTime - this.currentInterval.beg;
+        this.elapsed += this.currentInterval.elapsed;
+        this.runningElapsed = this.elapsed;
+        clearInterval(this.timer);
+        this.currentInterval = null;
+        // console.log("pause");
+    }
+
+    stop() {
+        if (!this.currentInterval) return;
+        this.pause();
+        this.end = Date.now();
+        this.callbacks = [];
+    }
+
+    del() {
+        if (!this.currentInterval) return;
+        if (this.timer) {
+            clearInterval(this.timer);
+            this.timer = null;
+        }
+        this.currentInterval = null;
+        this.callbacks = [];
+    }
+}
+
+export class TimeLogSessionHub {
+    sessions: { [key: string]: TimeLogSession };
+
+    constructor() {
+        this.sessions = {};
+    }
+
+    size() {
+        return Object.keys(this.sessions).length;
+    }
+
+    new(active: IActive) {
+        let session = new TimeLogSession(active);
+        let id = `${active.id}#${Date.now()}`;
+        this.sessions[id] = session;
+        session.runId = id;
+        return session;
+    }
+
+    del(session: TimeLogSession | string) {
+        let id = typeof session === "string" ? session : session.runId;
+        delete this.sessions[id];
+        console.log("Delete session", session);
+    }
+
+    pause() {
+        for (let id in this.sessions) {
+            this.sessions[id].pause();
+        }
+    }
+}
+
+export let sessionHub = new TimeLogSessionHub();
